Unify HyperView and Handler typeclass
fixed merge conflicts
