+++ src/Web/Hyperbole/Effect/Hyperbole.hs
@@ -1,10 +1,12 @@
 {-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE UndecidableInstances #-}
 
 module Web.Hyperbole.Effect.Hyperbole where
 
 import Control.Monad (join)
 import Data.Bifunctor (first)
 import Data.ByteString qualified as BS
+import Data.Kind (Constraint, Type)
 import Data.List qualified as List
 import Data.Maybe (isJust)
 import Data.String.Conversions
@@ -13,14 +15,16 @@ import Data.Text.Encoding (decodeUtf8, encodeUtf8)
 import Effectful
 import Effectful.Dispatch.Dynamic
 import Effectful.Error.Static
+import Effectful.Reader.Dynamic
 import Effectful.State.Static.Local
+import GHC.TypeLits hiding (Mod)
 import Web.FormUrlEncoded (Form, urlDecodeForm)
 import Web.HttpApiData (FromHttpApiData, ToHttpApiData (..), parseQueryParam)
 import Web.Hyperbole.Effect.Server
+import Web.Hyperbole.Handler.TypeList
 import Web.Hyperbole.HyperView
 import Web.Hyperbole.Route
 import Web.Hyperbole.Session as Session
-import Web.Hyperbole.View.Target (hyperUnsafe)
 import Web.View
 
 
@@ -110,18 +114,14 @@ formBody = do
   either (send . RespondEarly . Err . ErrParse) pure ef
 
 
-getEvent :: (HyperView id, Hyperbole :> es) => Eff es (Maybe (Event id (Action id)))
+getEvent :: (HyperView id es, Hyperbole :> es) => Eff es (Maybe (Event id (Action id)))
 getEvent = do
   q <- reqParams
-  pure $ parseEvent q
-
-
-parseEvent :: (HyperView id) => Query -> Maybe (Event id (Action id))
-parseEvent q = do
-  Event ti ta <- lookupEvent q
-  vid <- parseViewId ti
-  act <- parseAction ta
-  pure $ Event vid act
+  pure $ do
+    Event ti ta <- lookupEvent q
+    vid <- parseViewId ti
+    act <- parseAction ta
+    pure $ Event vid act
 
 
 lookupEvent :: Query -> Maybe (Event Text Text)
@@ -245,7 +245,7 @@ redirect = send . RespondEarly . Redirect
 
 
 -- | Respond with the given view, and stop execution
-respondEarly :: (Hyperbole :> es, HyperView id) => id -> View id () -> Eff es ()
+respondEarly :: (Hyperbole :> es, HyperView id es) => id -> View id () -> Eff es ()
 respondEarly i vw = do
   let vid = TargetViewId (toViewId i)
   let res = Response vid $ hyperUnsafe i vw
@@ -256,3 +256,170 @@ respondEarly i vw = do
 view :: (Hyperbole :> es) => View () () -> Eff es Response
 view vw = do
   pure $ Response (TargetViewId "") vw
+
+
+{- | HyperViews are interactive subsections of a 'Page'
+
+Create an instance with a unique view id type and a sum type describing the actions the HyperView supports. The View Id can contain context (a database id, for example)
+
+@
+data Message = Message Int
+  deriving (Generic, 'Param')
+
+data MessageAction
+  = Louder Text
+  | ClearMessage
+  deriving (Generic, 'Param')
+
+instance HyperView Message where
+  type Action Message = MessageAction
+@
+-}
+class (ViewId id, ViewAction (Action id)) => HyperView id es where
+  data Action id
+
+
+  type Require id :: [Type]
+  type Require id = '[]
+
+
+  handle :: (Hyperbole :> es) => Action id -> Eff (Reader id : es) (View id ())
+
+
+-- | The top-level view created by 'load'. Carries the views in its type to check that we handled all our views
+data Root (views :: [Type]) = Root
+  deriving (Show, Read, ViewId)
+
+
+instance HyperView (Root views) es where
+  data Action (Root views) = RootNone
+    deriving (Show, Read, ViewAction)
+  type Require (Root views) = views
+  handle _ = pure none
+
+
+type family ValidDescendents x :: [Type] where
+  ValidDescendents x = x : NextDescendents '[] '[x]
+
+
+type family NextDescendents (ex :: [Type]) (xs :: [Type]) where
+  NextDescendents _ '[] = '[]
+  NextDescendents ex (x ': xs) =
+    RemoveAll (x : ex) (Require x)
+      <++> NextDescendents ((x : ex) <++> Require x) (RemoveAll (x : ex) (Require x))
+      <++> NextDescendents (x : ex) (RemoveAll (x : ex) xs)
+
+
+type NotHandled id ctx (views :: [Type]) =
+  TypeError
+    ( 'Text "HyperView "
+        :<>: 'ShowType id
+        :<>: 'Text " not found in (Require "
+        :<>: 'ShowType ctx
+        :<>: 'Text ")"
+        :$$: 'Text "  "
+          :<>: 'ShowType views
+        :$$: 'Text "Try adding it to the HyperView instance:"
+        :$$: 'Text "  instance HyperView "
+          :<>: 'ShowType ctx
+          :<>: 'Text " where"
+        :$$: 'Text "    type Action "
+          :<>: 'ShowType ctx
+          :<>: 'Text " = "
+          :<>: ShowType (Action id)
+          :<>: 'Text ""
+        :$$: 'Text "    type Require "
+          :<>: 'ShowType ctx
+          :<>: 'Text " = ["
+          :<>: ShowType id
+          :<>: 'Text ", ...]"
+    )
+
+
+type NotDesc id ctx x cs =
+  TypeError
+    ( 'Text ""
+        :<>: 'ShowType x
+        :<>: 'Text ", a child of HyperView "
+        :<>: 'ShowType id
+        :<>: 'Text ", not handled by context "
+        :<>: 'ShowType ctx
+        :$$: ('Text " Require = " ':<>: 'ShowType cs)
+        -- ':$$: 'ShowType x
+        -- ':$$: 'ShowType cs
+    )
+
+
+type NotInPage x total =
+  TypeError
+    ( 'Text ""
+        :<>: 'ShowType x
+        :<>: 'Text " not included in: "
+        :$$: 'Text "  Page es "
+          :<>: ShowType total
+        :$$: 'Text "try expanding the page views to:"
+        :$$: 'Text "  Page es "
+          :<>: ShowType (x : total)
+          -- :$$: 'Text " " :<>: 'ShowType ctx :<>: 'Text " = " :<>: ShowType (Action id) :<>: 'Text ""
+          -- :$$: 'Text "    page :: (Hyperbole :> es) => Page es '[" :<>: 'ShowType ctx :<>: 'Text " = [" :<>: ShowType id :<>: 'Text ", ...]"
+    )
+
+
+type HyperViewHandled id ctx =
+  ( -- the id must be found in the children of the context
+    ElemOr id (Require ctx) (NotHandled id ctx (Require ctx))
+  , -- Make sure the descendents of id are in the context for the root page
+    CheckDescendents id ctx
+  )
+
+
+-- TODO: Report which view requires the missing one
+type family CheckDescendents id ctx :: Constraint where
+  CheckDescendents id (Root total) =
+    ( AllInPage (ValidDescendents id) total
+    )
+  CheckDescendents id ctx = ()
+
+
+type family AllInPage ids total :: Constraint where
+  AllInPage '[] _ = ()
+  AllInPage (x ': xs) total =
+    (ElemOr x total (NotInPage x total), AllInPage xs total)
+
+
+-- TODO: if I'm going to limit it, it's going to happen here
+-- AND all their children have to be there
+-- , All (Elem (Require ctx)) (Require id)
+hyper
+  :: forall id ctx
+   . (HyperViewHandled id ctx, ViewId id)
+  => id
+  -> View id ()
+  -> View ctx ()
+hyper = hyperUnsafe
+
+
+hyperUnsafe :: (ViewId id) => id -> View id () -> View ctx ()
+hyperUnsafe vid vw = do
+  el (att "id" (toViewId vid) . flexCol) $
+    addContext vid vw
+
+
+-- | Internal
+dataTarget :: (ViewId a) => a -> Mod
+dataTarget = att "data-target" . toViewId
+
+
+{- | Trigger actions for another view. They will update the view specified
+
+> otherView :: View OtherView ()
+> otherView = do
+>   el_ "This is not a message view"
+>   button OtherAction id "Do Something"
+>
+>   target (Message 2) $ do
+>     el_ "Now we can trigger a MessageAction which will update our Message HyperView, not this one"
+>     button ClearMessage id "Clear Message #2"
+-}
+target :: (HyperViewHandled id ctx) => id -> View id () -> View ctx ()
+target = addContext
