+++ src/Web/Hyperbole/Page.hs
@@ -5,7 +5,6 @@ import Effectful
 import Web.Hyperbole.Effect.Hyperbole
 import Web.Hyperbole.Effect.Server (Response)
 import Web.Hyperbole.Handler (RunHandlers, runLoad)
-import Web.Hyperbole.HyperView (Root)
 import Web.View (View)
