+++ src/Web/Hyperbole/View.hs
@@ -6,8 +6,8 @@ module Web.Hyperbole.View
   , module Web.View
   ) where
 
+import Web.Hyperbole.Effect.Hyperbole (hyper, target)
 import Web.Hyperbole.View.Element
 import Web.Hyperbole.View.Event
-import Web.Hyperbole.View.Target
 import Web.View hiding (Query, Segment, button, cssResetEmbed, form, input, label)
