+++ src/Web/Hyperbole/View/Event.hs
@@ -1,8 +1,8 @@
 module Web.Hyperbole.View.Event where
 
 import Data.Text (pack)
+import Web.Hyperbole.Effect.Hyperbole
 import Web.Hyperbole.HyperView
-import Web.Hyperbole.View.Target
 import Web.View (View, att, context, el, flexCol, hide, parent)
 
 
@@ -19,7 +19,7 @@ pollMessageView m = do
     'el_' ('text' m)
 @
 -}
-onLoad :: (HyperView id) => Action id -> DelayMs -> View id () -> View id ()
+onLoad :: (ViewId id, ViewAction (Action id)) => Action id -> DelayMs -> View id () -> View id ()
 onLoad a delay initContent = do
   c <- context
   el (att "data-on-load" (toAction a) . att "data-delay" (pack $ show delay) . dataTarget c) initContent
