+++ src/Web/Hyperbole/View/Target.hs
@@ -1,8 +1,6 @@
 module Web.Hyperbole.View.Target where
 
-import Web.Hyperbole.Handler.TypeList (HyperViewHandled)
-import Web.Hyperbole.HyperView
-import Web.View (Mod, View, addContext, att, el, flexCol)
+import Web.Hyperbole.Handler.TypeList ()
 
 
 {- | Embed HyperViews into the page, or nest them into other views
@@ -34,40 +32,3 @@ otherView = do
   button (Louder \"Hi\") id \"Louder\"
 @
 -}
-
--- TODO: if I'm going to limit it, it's going to happen here
--- AND all their children have to be there
--- , All (Elem (Require ctx)) (Require id)
-hyper
-  :: forall id ctx
-   . (HyperViewHandled id ctx)
-  => id
-  -> View id ()
-  -> View ctx ()
-hyper = hyperUnsafe
-
-
-hyperUnsafe :: (ViewId id) => id -> View id () -> View ctx ()
-hyperUnsafe vid vw = do
-  el (att "id" (toViewId vid) . flexCol) $
-    addContext vid vw
-
-
--- | Internal
-dataTarget :: (ViewId a) => a -> Mod
-dataTarget = att "data-target" . toViewId
-
-
-{- | Trigger actions for another view. They will update the view specified
-
-> otherView :: View OtherView ()
-> otherView = do
->   el_ "This is not a message view"
->   button OtherAction id "Do Something"
->
->   target (Message 2) $ do
->     el_ "Now we can trigger a MessageAction which will update our Message HyperView, not this one"
->     button ClearMessage id "Clear Message #2"
--}
-target :: (HyperView id) => id -> View id () -> View a ()
-target = addContext
